refactor(slack-bot): extract task-dedup.js from index.js (behavior-preserving)
